Add X session configuration
This should have been unnecessary, but the mainainter of the Debian
package for LightDM insisted that profile should not be sourced by
default (see the link in comment). Note that `$ANT_ARGS` was tested to
prevent sourcing twice, assuming that this variable will not be removed
from profile. This "include guard" was used because there doesn't seem
to be a reliable way to test whether current session has been started by
LightDM.