Add Rime configuration
Recipes are downloaded with Plum. Note that Plum isn't a complete recipe
manager; it adds and updates recipes, but doesn't remove any. To remove,
the recommended way is to clean up the Rime configuration directory, and
run Plum again. This is easier with "git clean".

The "symbols.yaml" isn't produced in Rime configuration root, nor the
"build" subfolder, so some online tutorials become stale. The original
file is available at:

https://github.com/rime/librime/blob/35db36e2a4e61e3dce9447833dbb2e579b165b33/data/minimal/symbols.yaml

The "trailing +" syntax for patch key specification is inspired
by

https://github.com/rime/rime-cantonese/blob/d3ecbe8e752c696e1dc0d64c610faaa8f855ce89/symbols_cantonese.yaml#L8

Unfortunately this syntax isn't mentioned in official documentation

https://github.com/rime/home/wiki/CustomizationGuide#定製指南

The "trailing @" syntax doesn't work because it's for YAML sequences,
while the key "punctuator/symbols" maps to a mapping, not a sequence.