Reorder the sections in the Zsh configuration
The chruby section is moved before the iTerm2 section so that, together
with the Zsh hook adjustment, Ruby version can be shown right after
changing directory. It was originally put after the local configuration
because `ls` was overriden there; this requirement went away with
0bfe4ded2990421732ef17ddceb11d875fa2df3e.

`chruby/auto.sh` is moved out of the workaround because it's not
affected by the issue.

NVM is also moved up, in case automatic invocation is added in future.