Recover the “optional source” function for POSIX
It wasn’t clear from the POSIX standard for dot whether a sourced script
counts as interactive, and should abort when the file is missing. In
reality, Zsh in POSIX emulation behaves the same as Dash, but Bash in
POSIX mode behaves differently. See the Stack Exchange question linked
in comment.